moving style duplications from templates to /common/style
