+++ sonar-project.properties
@@ -1,9 +1,9 @@
 sonar.projectKey=OpenCerts_opencerts-website
 sonar.projectName=OpenCerts_opencerts-website
 sonar.projectVersion=0.0.1
-sonar.sources=src
-sonar.tests=src
-sonar.test.inclusions=src/**/*.spec.js,src/**/*.spec.jsx,src/**/*.test.js,src/**/*.test.jsx
+sonar.sources=src/components/CertificateTemplates/tlds/sg/gov/ssg-wsg
+sonar.tests=src/components/CertificateTemplates/tlds/sg/gov/ssg-wsg
+# sonar.test.inclusions=src/**/*.spec.js,src/**/*.spec.jsx,src/**/*.test.js,src/**/*.test.jsx
 
 sonar.language=js
 sonar.profile=node
