+++ src/components/CertificateTemplates/tlds/sg/gov/ssg-wsg/SOA-IT-001/certificate.js
@@ -2,94 +2,12 @@ import { get } from "lodash";
 import { NICF_LOGO, IMG_LOGO, IMG_SEAL, IMG_SSGLOGO } from "../common";
 import { formatDate, formatCertID, getRecipientID } from "../common/functions";
 import fonts from "../common/fonts";
+import * as styles from "../common/style";
 
-export const fullWidthStyle = {
-  width: "100%",
-  height: "auto"
-};
-
-export const sealWidthStyle = {
-  width: "100%",
-  height: "auto"
-};
-
-export const signatureWidthStyle = {
-  width: "80%",
-  height: "auto"
-};
-
-export const printTextStyle = {
-  fontWeight: "500!important"
-};
-
-export const issuersTextStyle = {
-  fontWeight: "500!important",
-  fontSize: "24px"
-};
-
-export const transcriptTextStyle = {
-  fontWeight: "500!important",
-  fontSize: "24px",
-  marginBottom: "0px"
-};
-
-export const awardTextStyle = {
-  fontSize: "22px",
-  color: "rgb(197,41,155)",
-  textAlign: "left"
-};
-
-export const singaporeTextStyle = {
-  fontSize: "3rem"
-};
-
-export const nameTextStyle = {
-  fontSize: "2.3rem",
-  color: "rgb(197,41,155)",
-  wordBreak: "break-word",
-  fontWeight: "600"
-};
-
-export const recipientTextStyle = {
-  fontSize: "1.8rem",
-  marginBottom: "0px"
-};
-
-export const titleTextStyle = {
-  color: "rgb(30,93,200)",
-  fontSize: "3rem"
-};
-
-export const designationTextStyle = {
-  fontSize: "14px"
-};
-
-export const footerTextStyle = {
-  fontSize: "12px",
-  color: "rgb(51,0,144)",
-  marginTop: "15px"
-};
-
-export const footerAboutTextStyle = {
-  fontSize: "12px",
-  color: "rgb(51,0,144)"
-};
-
-export const certCodeStyle = {
-  fontSize: "12px",
-  color: "#ea649c",
-  display: "inline-block",
-  transform: "rotate(-90deg)"
-};
-
-export const footerLogoStyle = {
-  width: "auto",
-  height: "45px"
-};
 export const renderLogoWSQ = () => (
   <div className="row d-flex">
     <div className="col-lg-4 col-12">
-      <img style={fullWidthStyle} src={NICF_LOGO} />
+      <img style={styles.fullWidthStyle} src={NICF_LOGO} />
     </div>
     <div className="col-lg-6" />
   </div>
@@ -101,27 +19,27 @@ export const renderSignature = certificate => (
     style={{ marginTop: "8rem", marginBottom: "1rem" }}
   >
     <div className="col-lg-2 col-6">
-      <img style={sealWidthStyle} src={IMG_SEAL} />
+      <img style={styles.sealWidthStyle} src={IMG_SEAL} />
     </div>
     <div className="col-lg-10 col-12 row d-flex justify-content-center">
       <div className="col-lg-8">
         <div className="col-lg-3 col-12">
           <img
-            style={signatureWidthStyle}
+            style={styles.signatureWidthStyle}
             src={get(
               certificate,
               "additionalData.certSignatories[0].signature"
             )}
           />
         </div>
-        <div style={designationTextStyle} className="RobotoBold">
+        <div style={styles.designationTextStyle} className="RobotoBold">
           {get(certificate, "additionalData.certSignatories[0].name")},{" "}
           {get(certificate, "additionalData.certSignatories[0].position")}
         </div>
-        <div style={designationTextStyle} className="RobotoBold">
+        <div style={styles.designationTextStyle} className="RobotoBold">
           {get(certificate, "additionalData.certSignatories[0].organisation")}
         </div>
-        <div style={footerTextStyle} className="RobotoLight">
+        <div style={styles.footerTextStyle} className="RobotoLight">
           The training and assessment of the abovementioned learner are
           accredited
           <br />
@@ -131,13 +49,13 @@ export const renderSignature = certificate => (
       </div>
       <div className="col-lg-4 col-xs-12">
         <div style={{ marginBottom: "70px", marginTop: "60px" }}>
-          <p style={printTextStyle} className="RobotoRegular">
+          <p style={styles.printTextStyle} className="RobotoRegular">
             Cert No: {get(certificate, "additionalData.serialNum")}
           </p>
         </div>
       </div>
       <div className="col-lg-5 col-12">
-        <div style={footerAboutTextStyle} className="RobotoLight">
+        <div style={styles.footerAboutTextStyle} className="RobotoLight">
           <a href="www.ssg.gov.sg">www.ssg.gov.sg</a>
           <br />
           For verification of this certificate, please visit <br />
@@ -151,10 +69,10 @@ export const renderSignature = certificate => (
         style={{ alignItems: "center" }}
       >
         <div style={{ margin: "15px" }}>
-          <img style={footerLogoStyle} src={IMG_LOGO} />
-          <img style={footerLogoStyle} src={IMG_SSGLOGO} />
+          <img style={styles.footerLogoStyle} src={IMG_LOGO} />
+          <img style={styles.footerLogoStyle} src={IMG_SSGLOGO} />
         </div>
-        <div style={certCodeStyle}>
+        <div style={styles.certCodeStyle}>
           {get(certificate, "additionalData.certCode")}
         </div>
       </div>
@@ -165,22 +83,22 @@ export const renderSignature = certificate => (
 export const renderAwardText = certificate => (
   <div>
     <div className="d-flex" style={{ marginTop: "2rem" }}>
-      <p style={nameTextStyle} className="RobotoRegular">
+      <p style={styles.soaNameTextStyle} className="RobotoRegular">
         STATEMENT OF ATTAINMENT
       </p>
     </div>
     <div className="d-flex" style={{ marginTop: "3rem" }}>
-      <p style={awardTextStyle} className="RobotoMedium">
+      <p style={styles.awardTextStyle} className="RobotoMedium">
         is awarded to
       </p>
     </div>
     <div className="d-flex" style={{ marginTop: "1rem" }}>
-      <p style={recipientTextStyle} className="RobotoMedium">
+      <p style={styles.recipientTextStyle} className="RobotoMedium">
         {certificate.recipient.name}
       </p>
     </div>
     <div className="d-flex">
-      <p style={printTextStyle} className="RobotoMedium">
+      <p style={styles.printTextStyle} className="RobotoMedium">
         ID No: {getRecipientID(certificate.recipient)}
       </p>
     </div>
@@ -188,7 +106,7 @@ export const renderAwardText = certificate => (
       className="d-flex col-lg-6 col-12"
       style={{ marginTop: "1rem", marginBottom: "3rem", paddingLeft: "0px" }}
     >
-      <p style={awardTextStyle} className="RobotoMedium">
+      <p style={styles.awardTextStyle} className="RobotoMedium">
         for successful attainment of the following
         <br />
         industry approved competencies
@@ -196,13 +114,13 @@ export const renderAwardText = certificate => (
     </div>
     {certificate.transcript.map(item => (
       <div className="d-flex" key={item.courseCode}>
-        <p style={transcriptTextStyle} className="RobotoMedium">
+        <p style={styles.transcriptTextStyle} className="RobotoMedium">
           {item.courseCode} {item.name}
         </p>
       </div>
     ))}
     <div className="d-flex" style={{ marginTop: "3rem" }}>
-      <p style={issuersTextStyle} className="RobotoRegular">
+      <p style={styles.issuersTextStyle} className="RobotoRegular">
         at {certificate.additionalData.assessmentOrgName}
       </p>
     </div>
@@ -211,7 +129,7 @@ export const renderAwardText = certificate => (
 
 export const renderIssuingDate = certificate => (
   <div className="d-flex" style={{ marginTop: "1rem" }}>
-    <p style={issuersTextStyle} className="RobotoRegular">
+    <p style={styles.issuersTextStyle} className="RobotoRegular">
       {formatDate(certificate.attainmentDate)}
     </p>
   </div>
