+++ src/components/CertificateTemplates/tlds/sg/gov/ssg-wsg/common/style.js
@@ -1,0 +1,133 @@
+export const fullWidthStyle = {
+  width: "100%",
+  height: "auto"
+};
+
+export const sealWidthStyle = {
+  width: "100%",
+  height: "auto"
+};
+
+export const signatureWidthStyle = {
+  width: "80%",
+  height: "auto"
+};
+
+export const printTextStyle = {
+  fontWeight: "400",
+  color: "#000000"
+};
+
+export const issuersTextStyle = {
+  textAlign: "center",
+  fontSize: "24px"
+};
+
+export const transcriptTextStyle = {
+  fontSize: "24px",
+  marginBottom: "0px"
+};
+
+export const awardTextStyle = {
+  fontSize: "22px",
+  color: "rgb(197,41,155)",
+  textAlign: "left"
+};
+
+export const nameTextStyle = {
+  fontSize: "2.3rem",
+  textAlign: "left",
+  wordBreak: "break-word",
+  textTransform: "uppercase",
+  fontWeight: "600",
+  color: "#000000"
+};
+
+export const soaNameTextStyle = {
+  fontSize: "2.3rem",
+  color: "rgb(197,41,155)",
+  wordBreak: "break-word",
+  textAlign: "left"
+};
+
+export const transNameTextStyle = {
+  fontSize: "1.3rem"
+};
+
+export const recipientTextStyle = {
+  fontSize: "1.8rem",
+  textAlign: "left",
+  marginBottom: "0px"
+};
+
+export const titleTextStyle = {
+  color: "rgb(30,93,200)",
+  fontSize: "3rem",
+  textAlign: "center"
+};
+
+export const designationTextStyle = {
+  fontSize: "14px"
+};
+
+export const footerTextStyle = {
+  fontSize: "12px",
+  color: "rgb(51,0,144)",
+  marginTop: "15px"
+};
+
+export const footerLinkStyle = {
+  fontSize: "12px"
+};
+
+export const footerAboutTextStyle = {
+  fontSize: "12px",
+  color: "rgb(51,0,144)"
+};
+
+export const certCodeStyle = {
+  fontSize: "12px",
+  color: "#ea649c",
+  display: "inline-block",
+  transform: "rotate(-90deg)"
+};
+
+export const footerLogoStyle = {
+  width: "75%",
+  height: "auto"
+};
+
+export const transFooterLogoStyle = {
+  width: "80%",
+  height: "auto"
+};
+
+export const confermentTextStyle = {
+  textTransform: "uppercase"
+};
+
+export const qualHeaderStyle = {
+  fontWeight: "bold",
+  marginBottom: "0px"
+};
+
+export const qualTextStyle = {
+  fontSize: "2rem",
+  fontWeight: "bold",
+  textTransform: "uppercase"
+};
+
+export const singaporeTextStyle = {
+  fontSize: "3rem"
+};
+
+export const headerTextStyle = {
+  fontSize: "2.3rem",
+  color: "rgb(197, 41, 155)",
+  wordBreak: "break-word",
+  fontWeight: "600"
+};
+
+export const copyrightStyle = {
+  fontSize: "14px"
+};
